# openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/admin/properties": {
      "get": {
        "operationId": "AdminPropertyController_findAll",
        "summary": "List all properties (all statuses)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SOLD",
                "RENTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/properties/pending/list": {
      "get": {
        "operationId": "AdminPropertyController_findPending",
        "summary": "List properties pending review (paginated)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "enum": [
                "DRAFT",
                "PENDING",
                "APPROVED",
                "REJECTED",
                "SOLD",
                "RENTED"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/properties/{id}/approve": {
      "patch": {
        "operationId": "AdminPropertyController_approve",
        "summary": "Approve property (→ APPROVED)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "uuid-here",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/properties/{id}/reject": {
      "patch": {
        "operationId": "AdminPropertyController_reject",
        "summary": "Reject property (→ REJECTED)",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "example": "uuid-here",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectPropertyDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Properties"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
    "/owner/profile": {
      "get": {
        "operationId": "OwnerProfileController_getMyProfile",
        "summary": "Get my owner profile",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Owner Profile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/owner/profile/complete": {
      "post": {
        "operationId": "OwnerProfileController_completeProfile",
        "summary": "Complete owner profile (individual or company)",
        "description": "Submit or update KYC documents as images. Sets profileStatus → KYC_PENDING. Can resubmit while pending or after rejection.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "multipart/form-data": {
              "schema": {
                "$ref": "#/components/schemas/CompleteOwnerProfileMultipartDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Profile submitted"
          }
        },
        "tags": [
          "Owner Profile"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/owners/pending": {
      "get": {
        "operationId": "AdminOwnerController_listPending",
        "summary": "List owners pending KYC review (paginated)",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Page number (1-based)",
            "schema": {
              "default": 1,
              "example": 1,
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Items per page (max 100)",
            "schema": {
              "default": 20,
              "example": 20,
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Owner Review"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/owners/{userId}/approve": {
      "patch": {
        "operationId": "AdminOwnerController_approve",
        "summary": "Approve owner profile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Owner Review"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
    "/admin/owners/{userId}/reject": {
      "patch": {
        "operationId": "AdminOwnerController_reject",
        "summary": "Reject owner profile",
        "parameters": [
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RejectOwnerProfileDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Admin - Owner Review"
        ],
        "security": [
          {
            "access-token": []
          }
        ]
      }
    },
...
  },
  "info": {
    "title": "Aqar API",
    "description": "Rental Property Platform — Authentication, Authorization (RBAC), Properties & Bookings",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "in": "header"
      }
    },
    "schemas": {
...
      "RejectPropertyDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Images are unclear or listing details are incomplete"
          }
        },
        "required": [
          "reason"
        ]
      },
      "CompleteOwnerProfileMultipartDto": {
        "type": "object",
        "properties": {
          "ownerType": {
            "type": "string",
            "enum": [
              "INDIVIDUAL",
              "COMPANY"
            ],
            "example": "INDIVIDUAL"
          },
          "companyName": {
            "type": "string",
            "example": "شركة العقارات المتميزة"
          },
          "whatsapp": {
            "type": "string",
            "example": "+201098765432"
          },
          "phone": {
            "type": "string",
            "example": "+201234567890"
          },
          "email": {
            "type": "string",
            "example": "owner@example.com"
          },
          "address": {
            "type": "string",
            "example": "123 Main Street"
          },
          "city": {
            "type": "string",
            "example": "Cairo"
          },
          "area": {
            "type": "string",
            "example": "Nasr City"
          },
          "bio": {
            "type": "string",
            "example": "Experienced property owner with 10+ years."
          },
          "nationalId": {
            "type": "string",
            "format": "binary",
            "description": "National ID card image — required for INDIVIDUAL"
          },
          "taxNumber": {
            "type": "string",
            "format": "binary",
            "description": "Tax card image — required for COMPANY"
          },
          "commercialRegister": {
            "type": "string",
            "format": "binary",
            "description": "Commercial register image — required for COMPANY"
          }
        },
        "required": [
          "ownerType"
        ]
      },
      "RejectOwnerProfileDto": {
        "type": "object",
        "properties": {
          "reason": {
            "type": "string",
            "example": "Invalid commercial register document"
          }
        },
        "required": [
          "reason"
        ]
      }
    }
  }
}